Split DashboardScreen into new file
